share bulk of errorHandler between the servers
