+++ packages/central-server/app/middleware/errorHandler.js
@@ -1,9 +1,4 @@
-import { timingSafeEqual } from 'node:crypto';
-import config from 'config';
-import { BaseError as SequelizeError } from 'sequelize';
-import { convertDatabaseError } from '@tamanu/database';
-import { Problem } from '@tamanu/errors';
-import { log } from '@tamanu/shared/services/logging';
+import { errorHandlerProblem } from '@tamanu/shared/utils';
 
 // eslint-disable-next-line no-unused-vars
 export const buildErrorHandler = getResponse => (error, req, res, next) => {
@@ -13,40 +8,10 @@ export const buildErrorHandler = getResponse => (error, req, res, next) => {
     return;
   }
 
-  if (error instanceof SequelizeError) {
-    error = convertDatabaseError(error);
-  }
-
-  const exposeSensitive =
-    process.env.NODE_ENV !== 'production' ||
-    (typeof config.debugging.apiErrorsToken === 'string' &&
-      config.debugging.apiErrorsToken.length > 0 &&
-      timingSafeEqual(
-        Buffer.from(req.get('tamanu-debug') ?? ''),
-        Buffer.from(config.debugging.apiErrorsToken ?? ''),
-      ));
-  const problem = (
-    error instanceof Problem ? error : Problem.fromError(error)
-  ).excludeSensitiveFields(!exposeSensitive);
-
-  if (problem.status >= 500) {
-    log.error(`Error ${problem.status} (${problem.type}): `, error);
-  } else {
-    log.info(`Error ${problem.status} (${problem.type}): `, error);
-  }
+  const { problem, json } = errorHandlerProblem(error, req);
 
   res.set(problem.headers);
-  res.status(problem.status).send(getResponse(error, problem.toJSON()));
+  res.status(problem.status).send(getResponse(error, json));
 };
 
-export const defaultErrorHandler = buildErrorHandler((error, problem) => ({
-  // RFC 7807 Problem Details for HTTP APIs
-  ...problem,
-
-  // legacy error format
-  error: {
-    message: error.message,
-    name: error.name,
-    status: problem.status,
-  },
-}));
+export const defaultErrorHandler = buildErrorHandler((_error, defaultJson) => defaultJson);
