+++ packages/facility-server/app/middleware/errorHandler.js
@@ -1,35 +1,13 @@
-import { timingSafeEqual } from 'node:crypto';
-import config from 'config';
-import { BaseError as SequelizeError } from 'sequelize';
-import { convertDatabaseError } from '@tamanu/database';
-import { Problem } from '@tamanu/errors';
-import { log } from '@tamanu/shared/services/logging';
+import { errorHandlerProblem } from '@tamanu/shared/utils';
 
-// eslint-disable-next-line no-unused-vars
-export default function errorHandler(error, req, res, _) {
-  if (error instanceof SequelizeError) {
-    error = convertDatabaseError(error);
+export default function errorHandler(error, req, res, next) {
+  // see https://expressjs.com/en/guide/error-handling.html#the-default-error-handler
+  if (res.headersSent) {
+    next(error);
+    return;
   }
 
-  const exposeSensitive =
-    process.env.NODE_ENV !== 'production' ||
-    (typeof config.debugging.apiErrorsToken === 'string' &&
-      config.debugging.apiErrorsToken.length > 0 &&
-      timingSafeEqual(
-        Buffer.from(req.get('tamanu-debug') ?? ''),
-        Buffer.from(config.debugging.apiErrorsToken ?? ''),
-      ));
-  const problem = (
-    error instanceof Problem ? error : Problem.fromError(error)
-  ).excludeSensitiveFields(!exposeSensitive);
-
-  if (problem.type.includes('auth')) {
-    log.warn(`Error ${problem.status} (${problem.type}): ${error.message}`);
-  } else if (problem.status >= 500) {
-    log.error(`Error ${problem.status} (${problem.type}): `, error);
-  } else {
-    log.info(`Error ${problem.status} (${problem.type}): `, error);
-  }
+  const { problem, json } = errorHandlerProblem(error, req);
 
   // we're past the point of permission checking; this just
   // makes sure the error send doesn't get intercepted by the
@@ -39,11 +17,5 @@ export default function errorHandler(error, req, res, _) {
   }
 
   res.set(problem.headers);
-  res.status(problem.status).send({
-    ...problem.toJSON(),
-    error: {
-      message: error.message,
-      ...error,
-    },
-  });
+  res.status(problem.status).send(json);
 }
